Fixed a bug when computing offset in lib_seek functions

+++ Common/Libraries/audiolib/flac_lib.c
@@ -317,11 +317,11 @@ void flac_lib_seek(audio_lib_handle_t * hlib)
     BSP_AUDIO_OUT_Stop(CODEC_PDWN_HW);
     flac_lib_state = FLAC_LIB_FILL_READBUFF;
 
-    ofs = (hlib->seek_pos * hlib->fp->fsize) / 100;
+    ofs = ((uint64_t)hlib->seek_pos * hlib->fp->fsize) / 100;
 
     f_lseek(hlib->fp, ofs);
     
-    hlib->time.elapsed_time = (hlib->fp->fptr * 8) / codec->bitrate;   
+    hlib->time.elapsed_time = ((uint64_t)hlib->fp->fptr * 8) / codec->bitrate;   
 }
 
 void flac_lib_free(audio_lib_handle_t * hlib)
