+++ Common/Libraries/audiolib/remote_mp3_lib.c
@@ -297,9 +297,9 @@ void remote_mp3_lib_seek(audio_lib_handle_t * hlib)
     BSP_AUDIO_OUT_Stop(CODEC_PDWN_HW);
     mp3_lib_state = MP3_LIB_FILL_NETBUFF;
 
-    hlib->remote.fptr = (hlib->seek_pos * hlib->remote.fsize) / 100;
+    hlib->remote.fptr = ((uint64_t)hlib->seek_pos * hlib->remote.fsize) / 100;
 
-    hlib->time.elapsed_time = (hlib->remote.fptr * 8) / ((codec->info.bitrate / 1000) * 1024);    
+    hlib->time.elapsed_time = ((uint64_t)hlib->remote.fptr * 8) / ((codec->info.bitrate / 1000) * 1024);    
 }
 
 void remote_mp3_lib_free(audio_lib_handle_t * hlib)
