+++ Common/Libraries/audiolib/wav_lib.c
@@ -251,7 +251,7 @@ void wav_lib_seek(audio_lib_handle_t * hlib)
 
     payload_size = hlib->fp->fsize - sizeof(wav_info);
     nbuffers = payload_size / hlib->buffer->size;
-    frame_pos = (hlib->seek_pos * nbuffers) / 100;
+    frame_pos = ((uint64_t)hlib->seek_pos * nbuffers) / 100;
 
     ofs = sizeof(wav_info) + frame_pos * hlib->buffer->size;
